+++ CHANGELOG.md
@@ -76,6 +76,8 @@
 - UI: added fallback to front art setting for Art by TF (background setting), if no file was found.
 - UI: added new mask effect setting to background. See [here](https://hydrogenaudio.org/index.php/topic,112914.msg1084108.html#msg1084108).
 - UI: added new img counter setting to background while using Art by TF mode. See [here]https://hydrogenaudio.org/index.php/topic,116669.msg1083339.html#msg1083339).
+- UI: file/folder (by TF) modes now allow Microsoft MS-DOS wildcards at end of path. See [here](https://ss64.com/nt/syntax-wildcards.html).
+- UI: new art matching file sorting for file/folder (by TF) modes: creation date, explorer order and random order.
 - Exporting Playlist: added new Namida 'Export and convert' preset.
 - External integration: new callbacks.
 	* 'Playlist-Manager-SMP: switch enable panel'		-> { window: string[] }